+++ CHANGES.rst
@@ -52,6 +52,11 @@ astropy.stats
 astropy.table
 ^^^^^^^^^^^^^
 
+- Added ``units`` and ``descriptions`` keyword arguments to the Table object
+  initialization and ``Table.read()`` methods.  This allows directly setting
+  the ``unit`` and ``description`` for the table columns at the time of
+  creating or reading the table. [#9671]
+
 astropy.tests
 ^^^^^^^^^^^^^